Delete old examples and move new one to replace.
