+++ dotnet/samples/Concepts/Memory/VectorstoreLangchainInterop/RedisFactory.cs
@@ -10,7 +10,7 @@ namespace Memory.VectorstoreLangchainInterop;
 /// Contains a factory method that can be used to create a Redis vector store that is compatible with datasets ingested using Langchain.
 /// </summary>
 /// <remarks>
-/// This class is used with the <see cref="LangchainInterop"/> sample.
+/// This class is used with the <see cref="VectorStore_Langchain_Interop"/> sample.
 /// </remarks>
 public static class RedisFactory
 {
